core: dump TA call stack on TEE_Panic()
Provide a full call stack in case a TA panics. This is achieved by
having TEE_Panic() cause an abort at a specific virtual address defined
by CFG_TEE_PANIC_DEBUG_ADDR. When this special value is detected by the
abort handler, the diagnostic text is adapted, and the panic code is
preserved. Then, scripts/symbolize.py can be used to analyze the dump.
For instance (slightly edited to fit 72 characters):

 User TA TEE_Panic()
  fsr 0x00000005  ttbr0 0x0e07e06a  ttbr1 0x0e07806a  cidr 0x1
  cpu #0          cpsr 0x20000010
  r0 0x0000beef      r4 0x00102780    r8 0x00000000   r12 0x6dab895e
  r1 0x00000bad      r5 0x0012277f    r9 0x00000000    sp 0x001026e8
  r2 0x00136119      r6 0x00000000   r10 0x00000000    lr 0x00105361
  r3 0x00000001      r7 0x001026e8   r11 0x00000000    pc 0x001223b4
 Status of TA 5b9e0e40-2636-11e1-ad9e-0002a5d5c51b (0xe077c18) (active)
  arch: arm  load address: 0x103000  ctx-idr: 1
  stack: 0x100000 10240
  region 0: va 0x100000 pa 0xe41e000 size 0x3000 flags rw-
  region 1: va 0x103000 pa 0xe300000 size 0x2d000 flags r-x <snip>
  region 2: va 0x130000 pa 0xe32d000 size 0xb000 flags r-- <snip>
  region 3: va 0x13b000 pa 0xe338000 size 0xe6000 flags rw- .data .bss
  region 4: va 0 pa 0 size 0 flags ---
  region 5: va 0 pa 0 size 0 flags ---
  region 6: va 0 pa 0 size 0 flags ---
  region 7: va 0 pa 0 size 0 flags ---
 Call stack:
  0x001223b4 TEE_Panic at optee_os/lib/libutee/tee_api_panic_a32.S:38
  0x00105361 ta_entry_panic at optee_test/ta/os_test/os_test.c:739
  0x00105ca9 TA_InvokeCommandEntryPoint at
                                    optee_test/ta/os_test/ta_entry.c:89
  0x00122763 entry_invoke_command at
                      optee_os/lib/libutee/arch/arm/user_ta_entry.c:207
  0x001227bf __utee_entry at
                      optee_os/lib/libutee/arch/arm/user_ta_entry.c:235
user_ta_enter:452: tee_user_ta_enter: TA panicked with code 0xbeef

CFG_TEE_PANIC_DEBUG_ADDR obviously has to be an address that cannot be
mapped in the VA space of the TAs. The default value is 0xBAD, which is
below the TA stack region. When the variable is undefined or empty, the
feature is disabled and TEE_Panic() just calls utee_panic() ->
syscall_panic(), which provides only minimal debug information.

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>