core: review static mapping layout sequence
With this change, the small-page mapped areas are mapped around the
flat map areas to reuse the translation tables. If optee is memory
constrained running with the pager enables, the small page tables
use the map the flat map area will be use for other small page
mapped entries.

With this change, the virtual memory layout is design in 4 steps:
- define virtual location of flat mapped areas.
- define virtual location of small-page mapped areas around flat
  mapped areas (likely to already allocate xlat tables).
- define virtual location of pgdir mapped areas from bottom range
  to top range, skipping the already assigned virtual locations.

List of the constraints on virtual memory layout:
- optee_os includes some flat mapped sections.
- already leave the first pgdir entry unmapped for protection.
- Lpae mappings will locate userland virtual memory above core
  virtual memory.
- Non-lpae mappings must locate userland virtual memory below core
  virtual memory (ttbr0/ttbr1 referencing)

This change prepares mapping support of a coherent memory which will
add new constraints of the virtual memory layout.

Function init_smallpage_map() is used to assign virtual location
for small page regions inside a given virtual address range. This
clarifies sequence building the virtual mapping layout.

Signed-off-by: Etienne Carriere <etienne.carriere@linaro.org>