ci: .shippable.yml: fix commit range in global diff
When checking a pull request that has several commits, we want to run
checkpatch over the overall change. To do this, we need to identify the
first and last commits in the range and pass them to checkpatch.sh
which will then use them as `git rev-list first..last`.
The current code takes the first and last commits in
`git rev-list ${SHIPPABLE_COMMIT_RANGE}`. This is wrong, because
${SHIPPABLE_COMMIT_RANGE} is expressed with the three dots notation
(first...last), not the two dots one (first..last). git rev-list
interprets three dots as a symmetric difference, so we get all the
commits in the PR branch *and* in the target branch up to the common
point where the PR was created. We had a similar bug in the part that
checks each individual commit, fixed by commit 23b8cea77922 ("ci:
.shippable.yml: fix commit range").

Since we know the syntax of ${SHIPPABLE_COMMIT_RANGE}, we'd better
use a simpler approach and parse first...last instead.

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>