Makefile: Set ARCH := arm
The only value used for ARCH is arm at this point, not even aarch64, so
setting it immediately instead of only if absent to prevent build
errors should it be already defined to something else in the environment.

This can still be changed as an argument behind the make command of
course, but again, no other value is used or works other than arm.

Signed-off-by: Victor Chong <victor.chong@linaro.org>