[RFC] arm32: support for dynamically linked TAs
This commit goes half-way to adding support for dynamically linked
Trusted Applications. Therefore I'm sending it as a RFC.

Why do this? Several reasons.

1. Save space in the TA storage area. The OP-TEE core libraries
(libutee, libutils, libmpa) could very well be provided as shared
objects rather than object archives. They would be installed only once
in the TA storage. instead of being duplicated inside each TA.
2. Allow upgrade of library code without re-linking the TAs.
3. Pave the way to sharing code pages between TAs, thus reducing the
memory footprint of the TEE.

The ELF loader is updated as follows:

- Locate the dynamic section in the program headers (PT_DYNAMIC entry).
- Find the required external libraries by looking for DT_NEEDED entries
in the dynamic section. Libraries are .so files signed like TAs and
identified by a UUID so that the TA stores can be re-used unchanged.
- Load all the libraries.
- Process the dynamic relocations of type R_ARM_GLOB_DAT and
R_ARM_JUMP_SLOT by resolving symbols by name, in breadth first order.

What's left to be done:
- Map the library code and data into the user VA space (only the main
executable is mapped currently, so the TA crashes with a translation
fault as soon as an external symbol is referenced).
- Update the stack unwinding code. There is no longer a single EXIDX
exception table, but one per ELF file. Also, the crash dumps should
show the UUID of the libraries in the region mapping section.
- Update symbolize.py to handle the updated crash dumps.
- 64-bit support.

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>