scripts/symbolize.py: strip cwd from file paths by default
It is quite common to invoke symbolize.py from a top-level directory
which contains the source code of OP-TEE as well as of Trusted
Applications. In this case, it is convenient to strip the directory
from all the paths reported by the script.

Therefore, make this behavior the default. In other words, '-s $(pwd)'
is now implied. One can still obtain full paths by passing '-s' with no
argument.

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>
Acked-by: Etienne Carriere <etienne.carriere@linaro.org>