ci: .travis.yml: clone stable version (3.0.0)
Due to recent changes in manifest.git and build.git for buildroot
support,  it is not possible to build with GCC4.9 anymore. One has
to use the toolchain installed by "make toolchains".

Since we want to keep GCC4.9 for now [1], let's revert to an older,
stable version of the source tree, except for the optee_* projects
which we keep at the tip of their master branch.

[1] Commit 148ea708383e ("ci: .travis.yml: use GCC 4.9")

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>
Acked-by: Jens Wiklander <jens.wiklander@linaro.org>