github: Add CODEOWNERS file
CODEOWNERS file is a feature at GitHub that makes it possible to put
owners in certain parts of the code base. Linaro maintains the overall
code base, but for some platform specific implementations there are
platform maintainers. As a first try to leverage the GitHub features,
such as automate adding reviewers to pull requests we have put together
a CODEOWNERS file consisting of a catch-all for Linaro and then pairs of
platform specific code and the team of people who maintains that piece
of code.

This is one step of many in getting better control of pull requests,
issues etc at GitHub.

Signed-off-by: Joakim Bech <joakim.bech@linaro.org>