core: remove user_ta_elf::is_main
The 'is_main' boolean in struct user_ta_elf is used in a single
function: load_elf_from_store(). At this point we have another way of
determining if the ELF we're loading is the main executable or not.
Therefore, user_ta_elf::is_main is not really needed. Remove it.

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>
Reviewed-by: Jens Wiklander <jens.wiklander@linaro.org>